# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  source: 'source',
} as const;

export type UserSourceType = typeof UserSourceType[keyof typeof UserSourceType];


export const UserSourceType = {
  clinic: 'clinic',
  pharmacy: 'pharmacy',
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
  sourceName: string;
  sourceType: UserSourceType;
  locationName: string;
  lat: number;
  lng: number;
  createdAt: string;
}

export type RegisterInputSourceType = typeof RegisterInputSourceType[keyof typeof RegisterInputSourceType];


export const RegisterInputSourceType = {
  clinic: 'clinic',
  pharmacy: 'pharmacy',
} as const;

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  sourceName: string;
  sourceType: RegisterInputSourceType;
  /** @minLength 1 */
  locationName: string;
  lat: number;
  lng: number;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ReportSourceType = typeof ReportSourceType[keyof typeof ReportSourceType];


export const ReportSourceType = {
  clinic: 'clinic',
  pharmacy: 'pharmacy',
} as const;

export interface Report {
  id: number;
  submittedBy: number;
  sourceName: string;
  sourceType: ReportSourceType;
  locationName: string;
  lat: number;
  lng: number;
  symptom: string;
  patientCount: number;
  /** @nullable */
  medicineName: string | null;
  /** @nullable */
  unitsSold: number | null;
  reportDate: string;
  createdAt: string;
}

export interface ReportInput {
  /** @minLength 1 */
  symptom: string;
  /** @minimum 0 */
  patientCount: number;
  /** @nullable */
  medicineName?: string | null;
  /** @nullable */
  unitsSold?: number | null;
  reportDate: string;
}

export type OutbreakAlertSeverity = typeof OutbreakAlertSeverity[keyof typeof OutbreakAlertSeverity];


export const OutbreakAlertSeverity = {
  watch: 'watch',
  elevated: 'elevated',
  critical: 'critical',
} as const;

export interface OutbreakAlert {
  id: number;
  locationName: string;
  lat: number;
  lng: number;
  symptom: string;
  zScore: number;
  alertDate: string;
  severity: OutbreakAlertSeverity;
  createdAt: string;
}

export interface AdminOverview {
  totalReports: number;
  reportsLast7Days: number;
  activeAlerts: number;
  registeredSources: number;
  locationsMonitored: number;
  /** @nullable */
  lastDetectionAt: string | null;
}

export interface DetectionResult {
  alertsCreated: number;
  detectedAt: string;
}

export interface HeatmapPoint {
  lat: number;
  lng: number;
  intensity: number;
  locationName: string;
}

export interface TrendPoint {
  date: string;
  count: number;
  isAnomaly: boolean;
  /** @nullable */
  zScore: number | null;
}

export type ListAdminReportsParams = {
location?: string;
symptom?: string;
from?: string;
to?: string;
};

export type GetTrendParams = {
location: string;
symptom: string;
};