# lib/api-zod/src/generated/types/adminOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminOverview {
  totalReports: number;
  reportsLast7Days: number;
  activeAlerts: number;
  registeredSources: number;
  locationsMonitored: number;
  /** @nullable */
  lastDetectionAt: Date | null;
}