# lib/api-zod/src/generated/types/detectionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionResult {
  alertsCreated: number;
  detectedAt: Date;
}