# lib/api-zod/src/generated/types/heatmapPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapPoint {
  lat: number;
  lng: number;
  intensity: number;
  locationName: string;
}