# lib/api-zod/src/generated/types/outbreakAlert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OutbreakAlertSeverity } from './outbreakAlertSeverity';

export interface OutbreakAlert {
  id: number;
  locationName: string;
  lat: number;
  lng: number;
  symptom: string;
  zScore: number;
  alertDate: Date;
  severity: OutbreakAlertSeverity;
  createdAt: Date;
}