# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputSourceType } from './registerInputSourceType';

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  sourceName: string;
  sourceType: RegisterInputSourceType;
  /** @minLength 1 */
  locationName: string;
  lat: number;
  lng: number;
}