# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  /** @minLength 1 */
  symptom: string;
  /** @minimum 0 */
  patientCount: number;
  /** @nullable */
  medicineName?: string | null;
  /** @nullable */
  unitsSold?: number | null;
  reportDate: Date;
}